Move class caching and Rtti -> Class lookup to scalanative.runtime
